Add the trait ReadB and Impl it for Regs{W,B}
It'd be nice to have two functions per operation - foo() and foow().
In order to achieve this, we need to pass foo() a type with a Trait ReadB -
this is to ensure that the same function can be used for 8 bit register
contents AND the 8 bits read from memory pointed to by a 16 bit register.

i.e

foo(A)
foo(AF) where AF is

   AF        Addr    Val
            0x1233   0x00
[0x1234] -> 0x1234   0xFF
            0x1235   0x00